==> script/Analysis 2023 for Forestry/5-subset data for Forestry.Rmd <==
```{r include=FALSE}
#樣點資料檢核
#---- load library

library(tidyverse)
library(readxl)
library(writexl)
library(sf)
library(here)
#------------
```

欄位名中英文對照

```{r}
colname_en <- c('Office','Station','Site_N','Name',
                'Year','Month','Day',
                'Survey','Surveyor','Point',
                'TWD97_X','TWD97_Y','Hour','Minute',
                'Macaca_sur','Macaca_dist','Macaca_voice','Habitat', 'Note')

colname_ch <- c('分署','工作站','樣區編號','樣區名稱',
                '年','月','日',
                '旅次','調查者','樣點編號',
                'X座標\\(TWD97\\)','Y座標\\(TWD97\\)','時','分',
                '數量','距離','叫聲','棲地類型\\(主要\\)','備註')
```



```{r}
# path <- "./data/clean/Forestry/full_combind/"
# 
# DF <- 
# list.files(here(path), full.names = T) %>% 
#   map(read_excel, sheet = "Data", col_types ="text") %>%
#   map(function(x) {
#     x %>%
#        setNames(str_replace_all(names(.),setNames(colname_en,colname_ch))) %>%
#     setNames(str_replace_all(names(x),
#                              c("distance" = "Distance",
#                                "join_TypeName"= "TypeName"))
#              )
#     }) %>%
#   
#   bind_rows()
```


- Macaca_sur = 1 =   猴群
- Macaca_sur = 0 =   非猴群(無、孤猴)
- Macaca_sur = NA =  沒調查


```{r}

M.data <- read_excel(here("./data/clean/Forestry/full_combind/full_combind_Forestrydata_2023_V1.xlsx"), sheet = "Data")%>%
       setNames(str_replace_all(names(.),setNames(colname_en,colname_ch))) %>%
    setNames(str_replace_all(names(.),
                             c("distance" = "Distance",
                               "join_TypeName"= "TypeName"))
             )%>% 
  select(-Macaca_voice, -Habitat ) %>% 
  mutate(Macaca_sur = ifelse(Macaca_dist %in% "C" , 0, Macaca_sur))



# 
# M.data <-
#   DF%>% 
#   select(-Macaca_voice, -Habitat ) %>% 
#   mutate(Macaca_sur = ifelse(Macaca_dist %in% "C" , 0, Macaca_sur))
```

```{r}
M.0<- M.data %>%
  filter(analysis %in% "Y") %>% 
  filter(Macaca_sur %in% 1)  #猴群

M.2<- M.0 %>%  #找出同一旅次同一樣區有猴群大於2者
  filter(analysis %in% "Y") %>% 
  group_by(Year, Survey, Site_N) %>% 
  summarise(N = n()) %>% 
  filter(N>1)
```

```{r}
M.abc <- M.2 %>%
  left_join(M.0, by = c("Year", "Survey","Site_N")) %>%   #幫M.2這樣區樣點加回座標等資訊
  mutate(Y_S_S = paste0(Year, "_", Survey, "_",Site_N)) %>% 
  st_as_sf(., coords = c("TWD97_X", "TWD97_Y"), crs = 3826) %>% 
  
  split(., .$Y_S_S) %>% 
  lapply(., function(x){

    dist <-   
      st_distance(x$geometry)    #算點與點間的距離
    
      colnames(dist) <- x$Point         #加colnames
      
      dist %>% 
      as.data.frame() %>%   #轉回長表格
      setNames(., x$Point) %>% 
      add_column(Base_point = x$Point) %>% 
      reshape2::melt(id.vars = "Base_point",
                     variable.name = "Nearest_point",
                     value.name = "distance")%>%  
      filter(!distance %in% 0) %>% 
        arrange(Base_point)
    }) %>% 
  bind_rows(.id = "Y_S_S") %>% 
  separate("Y_S_S", c("Year", "Survey","Site_N"), "_") %>% 
  mutate(distance = as.numeric(distance)) %>% 
  mutate(Nearest_point = Nearest_point %>% as.character() %>% as.numeric()) %>% 
  mutate(Base_point = as.numeric(Base_point))
```

```{r}
#M.abc %>% write_xlsx(.,"D:/待處理工作夾(做完要歸檔)/重複記錄.xlsx")  
#重複記錄用人工卻認吧，依樣區有3個以上的重複記錄就很麻煩，程式規則也寫很麻煩

#1. 同一樣區同一旅次300m內只有一群
#2. 如果只有兩群以上，建議看一下有猴群的樣點位置，合理的推測一下要刪掉哪一個樣點的猴群，尤其是相鄰樣點會有隔壁樣點也聽到的情形，ex:連續3樣點都有猴群紀錄，有可能的情形是中間那點有猴群，旁邊兩個樣點都聽到。
```

```{r}
remove.data <-   #重複記錄的列表
  M.abc %>%  
  filter(distance < 300)%>% 
  mutate( Y_S_S = paste0(Year, "_", Survey, "_",Site_N)) %>% 
  split(., .$Y_S_S) %>% 
  lapply(., function(x){
    
    if(nrow(x)==2){
      x <- x[2,]  #要刪除第2點
    }else{
      x  #2點以上，留下來再討論，優先考量留下距離段近的，較遠的刪除
    }
  }) %>% 
  bind_rows %>% 
  # 手動
  filter(!( Year %in% 2020 & Survey %in% 1 & Site_N %in% "MA-H31-04" & Base_point %in% c(2, 4))) %>% 
  filter(!( Year %in% 2020 & Survey %in% 1 & Site_N %in% "MA-H32-06" & Base_point %in% c(2, 4))) %>% 
  filter(!( Year %in% 2020 & Survey %in% 2 & Site_N %in% "MA-E21-11" & Base_point %in% c(3, 5))) %>%
  filter(!( Year %in% 2021 & Survey %in% 1 & Site_N %in% "MA-A02-06" & Base_point %in% c(1, 3))) %>%
  filter(!( Year %in% 2021 & Survey %in% 1 & Site_N %in% "MA-A05-01" & Base_point %in% c(4, 6))) %>%
  filter(!( Year %in% 2021 & Survey %in% 1 & Site_N %in% "MB-E20-04" & Base_point %in% c(3, 5))) %>%
  filter(!( Year %in% 2021 & Survey %in% 2 & Site_N %in% "MA-A05-01" & Base_point %in% c(2, 4))) %>%
  filter(!( Year %in% 2021 & Survey %in% 2 & Site_N %in% "MA-E21-10" & Base_point %in% c(1, 4))) %>%
  filter(!( Year %in% 2022 & Survey %in% 1 & Site_N %in% "MA-F26-05" & Base_point %in% c(2, 4, 6))) %>%
  filter(!( Year %in% 2022 & Survey %in% 1 & Site_N %in% "MA-H32-10" & Base_point %in% c(2, 3))) %>%
  filter(!( Year %in% 2022 & Survey %in% 2 & Site_N %in% "MA-H34-09" & Base_point %in% c(2, 6))) %>%
  filter(!( Year %in% 2023 & Survey %in% 2 & Site_N %in% "MA-E22-07" & Base_point %in% c(2, 4))) %>%
  filter(!( Year %in% 2023 & Survey %in% 2 & Site_N %in% "MA-F26-01" & Base_point %in% c(6))) %>%
  
  
  
  
  # ~手動
  select( -Nearest_point, -distance, -Y_S_S) %>% 
  unique()%>% 
  setNames(., c("Year", "Survey","Site_N", "Point")) %>% 
  mutate(Point = as.character(Point))
  
```



```{r eval=FALSE, include=FALSE}
M.data.1 <-
  #remove 重複的猴群
  M.data %>% 
  semi_join(remove.data) %>% 
  mutate(Macaca_sur = 0)%>% 
  bind_rows(.,
            anti_join(M.data, remove.data) ) %>% 
  arrange(Year, Survey, Site_N, Point)   
  
  #~~remove 重複的猴群
```

```{r}
View(M.data.1 )
```

```{r}
M.data %>%  #移除重複前
  filter(Macaca_sur %in% 1) %>% 
  group_by(Year, Survey, Site_N) %>% 
  summarise(N = n()) %>%
  .$N %>% 
  table
```

```{r}
M.data.1 %>%   #移除重複後  #樣區數應該不會變
  filter(Macaca_sur %in% 1) %>% 
  group_by(Year, Survey, Site_N) %>% 
  summarise(N = n()) %>% 
  .$N %>% 
  table

# for analysis2022
```

```{r}
M.data.1 <- 
M.data.1 %>% 
  mutate(TypeName.1 = ordered(TypeName.1,c("闊葉林", "針葉林","混淆林","竹林","非森林"))) %>% 
  mutate(julian.D = ISOdatetime(Year, Month, Day, Hour, Minute, sec = 0) %>%
           as.POSIXlt(format = "%yyyy%mm%dd") %>%
           format(., "%j") %>% as.numeric())
```

```{r}
M.data.2 <-
  M.data.1 %>% 
  mutate(analysis = ifelse(analysis == "Y" &( Month < 3 | Month > 6), 
                           "month", analysis)) %>%   #刪除調查季(包含緩衝期)以外的資料
  mutate(analysis = ifelse(analysis == "Y" &  Altitude<50, 
                           "Y(50m)", analysis)) %>%  
  mutate(analysis = ifelse(analysis == "Y" &  TypeName.1 %in% "非森林", 
                           "Y(NotForest)", analysis))
```

```{r}
point_summary <- 
M.data.2 %>% 
  summarise(
    '1 收到' = n(),
    '2-1 6min' = str_subset(analysis, "6min") %>% length,
    '2-2 Toolate' = str_subset(analysis, "Toolate") %>% length,
    '2-3 7day' = str_subset(analysis, "7day") %>% length,
    '2-4 locate' = str_subset(analysis, "locate") %>% length,
    '2-5 不在3~6月內' = str_subset(analysis, "month") %>% length,
    'end of 2' = str_subset(analysis, "locate|6min|Toolate|7day|month", negate = T) %>% length,
    '3 低於50m' = str_subset(analysis, "50m") %>% length,
    '4 非森林' = str_subset(analysis, "NotForest") %>% length,
    '5 分析用' = str_subset(analysis, "^Y$") %>% length
  ) %>% 
  
  reshape2::melt(variable.name = "項目",
                 value.name = "樣點次")
```

```{r}
NOTE <- data.frame(
  說明 = c(
    "6min, Toolate, 7day, locate = 不符合調查規範",
    "month = 符合調查規範，但月份不在3~6月",
    "50m = 符合調查規範述，月份在3~6月，但海拔<50m",
    "NotForest = 符合調查規範述，月份在3~6月，海拔>=50m，但非森林",
    "Y = 符合調查規範述，月份在3~6月，海拔>=50m，只有森林"
         ))
```

```{r}
write_xlsx( list('NOTE' = NOTE,
                 'Point_Summary' = point_summary,
                 'Data' = M.data.2), 
            paste0(here("./data/clean/Forestry/for analysis/for analysis Forestrydata_2023_V1.xlsx")))



```
