==> script/Analysis 2023 for Forestry/9-rtrim.Rmd <==
```{r include=FALSE}
library(tidyverse)
library(rtrim)
library(openxlsx)
library(rtrim)
library(readxl)
library(here)
library(flextable)
library(RColorBrewer) 
library(sf)


```


```{r include=FALSE}
ecoregion3 <- 
  st_read(
    "D:/R/test/BBS site name_gis_data/ecoregion3HJShiu2003_97.shp") %>% 
  st_transform( crs=3826) %>% 
  dplyr::select(ZONE41)

```

```{r include=FALSE}

weight.o <- read_excel("D:/R/test/bbs_handover_temp_v20190123 - WT/weight.xlsx")  %>% 
  setNames(., c("Name", "cov1", "Area", "Site_n","prob_Site", "prob_Area",  "weight")) %>% 
  filter(Name %in% c('北部', '東部', '西部', '中海拔', '高海拔')) %>% 
  dplyr::select(Name, Area) %>% 
  mutate(Region = case_when(
    Name == '北部' ~ "North",
    Name == '東部' ~ "East",
    Name == '西部' ~ "West",
    TRUE ~ "Mountain",
  )) %>% 
  group_by(Region) %>% 
  summarise(Area = sum(Area)) %>% 
  mutate(prob_Area = Area/ sum(Area))


```



```{r include=FALSE}
#Original data---- 

M.data <- 
  list.files(here("./data/clean/Forestry/for analysis/"),
             full.names = T,pattern = "xlsx$|xls$") %>% 
  lapply(., read_excel, sheet="Data", col_types = "text") %>% 
  bind_rows() %>% 
  
  filter(analysis %in% "Y")%>% 
  
  st_as_sf(., coords = c("TWD97_X", "TWD97_Y"),crs=3826) %>% 
  st_join(., ecoregion3) %>% 
  st_drop_geometry() %>% 
  mutate(ZONE41 = as.character(ZONE41)) %>% 
  mutate(Region = case_when(
    as.numeric(Altitude) <1000 & ZONE41  %in% "1" ~ "North",
    as.numeric(Altitude) <1000 & ZONE41  %in% "2" ~ "East",
    as.numeric(Altitude) <1000 & ZONE41  %in% "3" ~ "West",
    as.numeric(Altitude) >= 1000 ~ "Mountain"
  ))
```


```{r include=FALSE}
weight <- 
  M.data %>% 
  mutate(SP = paste0(Site_N,"-",Point)) %>% 
  dplyr::select(Year, SP, Survey, Region) %>% 
  unique %>%
  reshape2::dcast(SP + Region ~ Year, length) %>% 
  reshape2::melt(id = 1:2, variable.name = "Year", value.name = "point_n") %>% 
  group_by(Region) %>% 
  mutate(SP_n = SP %>% unique %>% length())%>% 
  left_join(weight.o)%>% 
  mutate(point_n = ifelse(point_n %in% 0, 1, point_n) ) %>% 
  mutate(weight = 1000* prob_Area / SP_n /point_n) %>% 
  dplyr::select(Year, SP, Region, weight)
#weight := prob_Area / site_n /調查次數  
 

```

```{r include=FALSE}
df <- 
  M.data %>% 
  mutate(SP = paste0(Site_N,"-",Point)) %>%
  group_by(Year, SP, Region) %>% 
  summarise(number = Macaca_sur %>% as.numeric %>% sum(na.rm = T)) %>% 
  ungroup %>% 
  reshape2::dcast(SP +Region ~ Year, value.var = "number") %>% 
  reshape2::melt(id = 1:2, variable.name = "Year", value.name = "number")%>% 
  left_join(weight) %>% 
  
  group_by(SP, Region) %>% 
  mutate(N = sum(number, na.rm = T)) %>% 
  filter(!(N %in% 0)) %>% 
  dplyr::select(-N) %>% 
  
  mutate(Year = Year %>% as.character()%>% as.integer())%>% 
  mutate(number = number %>% as.character()%>% as.integer()) %>% 
  as.data.frame()
  

```


```{r echo=FALSE}
m1 <- trim(
  number ~ SP + Year + Region ,
  weights = "weight",df,
  model =  2,
  changepoints = "all",
  overdisp = F,
  serialcor = F, 
  autodelete = T, 
  stepwise = F)
```


```{r echo=FALSE}
 gof(m1) 

```

wald


Wald test for significance of covariates
```{r echo=FALSE, message=FALSE, warning=FALSE, paged.print=FALSE}
wald(m1) %>% 
  .$covar %>% 
  print %>% 
  flextable(.)
```

Wald test for significance of changes in slope
```{r echo=FALSE, message=FALSE, warning=FALSE, paged.print=FALSE}
wald(m1) %>% print %>% flextable(.)
```



Index  
```{r echo=FALSE}
index(m1, "imputed", covars = F) %>%
  plot(., pct =T, axes = F)
axis(1, at = c(2020:2023))
axis(2)
box()

```


```{r include=FALSE}
#covars = T

index(m1, "imputed", covars = T) %>% plot
```

Overall
```{r echo=FALSE, message=FALSE, warning=FALSE}
overall(m1,"imputed") %>% 
  print(.) %>% flextable()
```

overall  
```{r eval=FALSE, include=FALSE}
plot(overall(m1, "imputed"), axes = F)
axis(1, at = 2015:2023, cex.axis=1.5)
axis(2, cex.axis=1.5)
box()
title( "imputed" )
```

heatmap
```{r eval=FALSE, include=FALSE}
heatmap(m1, "imputed", cex.axis=1.5) 
title( "imputed" )
```

index

```{r message=FALSE, warning=FALSE, include=FALSE}
library(showtext)
library(ggprism)

showtext_auto()
font_add("Microsoft JhengHei", "msjh.ttc")

theme_bbs <-  
  theme_bw() +
  theme(
    plot.background = element_blank(),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_blank(),
    axis.title.x = element_blank(),
    axis.title.y = element_blank(),
    axis.text.x = element_text(size = 20,angle = 0, hjust = 0.5),
    text = element_text(size = 30, family = "Microsoft JhengHei"),
    plot.title = element_text(size = 30, hjust = 0.5),
    axis.line = element_line(color = "black", size = .1),
    axis.ticks = element_line(color = "black", size = .1)
  ) 

```


```{r include=FALSE}
idx <- index(m1, "imputed", covars = T)  %>% 
  dplyr::select(covariate, category, time, imputed = imputed, se_imp = se_imp)

p <- 
ggplot(idx[idx$covariate %in% "Overall",], aes(x = time, y= imputed*100 ))+
  geom_hline(yintercept=100, linetype = 2, linewidth = 0.3)+
  geom_errorbar(
    aes(x=time,
        ymax = 100*(imputed + se_imp),
        ymin = 100*(imputed - se_imp)),
    width=.03, colour = '#d6ce93', size = .3)+
  geom_line(data = idx[idx$covariate %in% "Overall",],
            aes(x = time, y= imputed*100 ),
            colour = '#d6ce93', linewidth = 0.6, linetype = 1)+
  geom_point(colour = '#d6ce93', size = 2.5, shape = 21,
             stroke = 0.6, fill = '#ef4926')+
          
      #    labs(title = fig_title) +
          
          expand_limits(y = 0) +
          scale_x_continuous(
           breaks = c(seq(2020, 2023,1))#,
   #          labels = c(seq(2020, 2022,1))
          )+
          scale_y_continuous()+
          theme_bbs+
          coord_cartesian(clip = "off") + 
          annotation_ticks(sides = "b", type = "both",size = 0.2,
                           outside = T,
                           tick.length = unit(1, "mm"),
                           minor.length = unit(0.5, "mm"))

  ggsave(filename = paste0("trim_index_MACACA.png"),p, device ="png",
               path =here("./Report of Foresty_20231026"),
               width = 5, height = 5, units = "cm",dpi = 300)

```


LCL、UCL、idx_2023指標變化值
```{r echo=FALSE}

Overall.mul <- overall(m1, "imputed")$slope$mul
Overall.se_mul <- overall(m1, "imputed")$slope$se_mul

round( 100*(Overall.mul-1.96*Overall.se_mul) ,1) #`LCL`
round( 100*(Overall.mul+1.96*Overall.se_mul) ,1) #`UCL`

idx_2023 <- idx[idx$covariate %in% "Overall" & idx$time %in% "2023",]$imputed

round( 100*(idx_2023 -1) ,0)  # idx_2022指標變化值

```
